// src/api/schemas/commentOut.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { AuthorSchema } from './authorSchema';
import type { CommentOutLink } from './commentOutLink';
import type { CommentOutMedia } from './commentOutMedia';
import type { CommentOutParentId } from './commentOutParentId';
import type { CommentOutTruncatedAtDepth } from './commentOutTruncatedAtDepth';
import type { CommentOutUserReactions } from './commentOutUserReactions';

/**
 * Comment output schema with nested replies limited to depth 6
and default_collapsed support
 */
export interface CommentOut {
  id: number;
  content: string;
  parent_id?: CommentOutParentId;
  created_at: string;
  updated_at: string;
  is_edited: boolean;
  like_count: number;
  reply_count: number;
  depth: number;
  replies?: CommentOut[];
  user_reactions?: CommentOutUserReactions;
  is_author?: boolean;
  media?: CommentOutMedia;
  link?: CommentOutLink;
  is_deleted?: boolean;
  author_info: AuthorSchema;
  has_more_replies?: boolean;
  truncated_at_depth?: CommentOutTruncatedAtDepth;
  default_collapsed?: boolean;
}